+++ changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [unreleased]
 
 - Renamed NewSvg to NewSVG
+- Moved type shape.Style to package draw
 
 ## [0.12.0] 2020-09-09
 ### Changed