+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Moved type shape.Style to package draw
 - Renamed NewSvg to NewSVG
 
 ## [0.12.0] 2020-09-09